pyproject.toml
[tool.poetry]
name = "py-irt"
version = "0.1.1"
description = "Bayesian IRT models in Python"
authors = ["John P. Lalor <john.lalor@nd.edu>"]
license = "MIT License"

[tool.poetry.dependencies]
python = ">=3.7,<3.10"
numpy = "^1.20.3"
pandas = "^1.2.4"
pyro-ppl = "^1.6.0"
rich = "^9.3.0"
scipy = "^1.6.3"
typer = "^0.3.2"
pydantic = "^1.8.2"
pedroai = "^0.1.15"

[tool.poetry.dev-dependencies]
codecov = "^2.1.11"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
pylint = "^2.8.3"
black = "^21.5b2"
isort = "^5.8.0"
ipython = "^7.24.1"

[tool.black]
line-length = 100

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
